requirements.txt
bleach==5.0.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
commonmark==0.9.1
cryptography==37.0.2
docopt==0.6.2
docutils==0.18.1
gnureadline==8.0.0
idna==3.3
importlib-metadata==4.11.4
inquirerpy==0.3.4
jeepney==0.8.0
keyring==23.6.0
pfzy==0.3.4
pkginfo==1.8.3
prompt-toolkit==3.0.30
pycparser==2.21
Pygments==2.12.0
python-dotenv==0.20.0
readme-renderer==35.0
requests==2.28.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
SecretStorage==3.3.2
six==1.16.0
twine==4.0.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.8.0
